requirements.txt
# Web3 & Blockchain
web3>=6.0.0

# AI Frameworks
langgraph>=0.0.10
langchain-groq>=0.1.0
langchain-core>=0.2.0

# Infrastructure & Utilities
python-dotenv>=1.0.0